Increase Test Coverage (#145)
* feat: Increase test coverage to over 90%

This commit increases the test coverage for the entire application to over 90%.

- Added tests for repositories, adapters, domains, controllers, and use cases that were below the 90% threshold.
- Refactored existing tests to cover more edge cases and error conditions.
- Added a missing authentication check in the JobController.
- Refactored the update method in the base Repository to handle entities without an ID more cleanly.

* feat: Increase test coverage to over 90%

This commit increases the test coverage for the entire application to over 90%.

- Added tests for repositories, adapters, domains, controllers, and use cases that were below the 90% threshold.
- Refactored existing tests to cover more edge cases and error conditions.
- Added a missing authentication check in the JobController.
- Refactored the update method in the base Repository to handle entities without an ID more cleanly.
- Fixed linting errors.

---------

+++ src/controllers/JobController.ts
@@ -98,6 +98,9 @@ export default class JobController {
     };
 
     public deleteJob = async (req: Request<JobParams>, res: Response<{ msg: string }>) => {
+        if (!req.user) {
+            throw new UnauthenticatedError("Authentication Invalid");
+        }
         const useCase = new DeleteJobUseCase(
             this.jobRepository,
             this.userRepository
